requirements.txt
pymysql
flatten_json
pyyaml
tabulate
